requirements.txt
discord.py
anthropic
pytz
